backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
google-genai
pymongo
pydub 